requirements.txt
streamlit==1.48.1
pytest==8.4.1
langchain==0.3.27
langchain-openai==0.3.30
langchain-core==0.3.74
python-dotenv==1.1.1
pydantic==2.11.7
langgraph==0.6.4
